+++ components/views/settings.ts
@@ -1,5 +1,5 @@
 export default {
   // currently used for custom sensors such as Met One;
-  // todo(nc): this should be removed when we have manifests
+  // todo(nc): this should be removed when we have nodeMetas
   mdpNodes: ['W038', 'W01D', 'W01F', 'V008']
 }